Created KMeansCommons file for common code.  Created trait holding functions common to KMeans and KMeansMiniBatch objects.
